requirements.txt
# DAST Scan Automation - Python Dependencies
# Install with: pip install -r requirements.txt

PyYAML>=6.0
google-cloud-storage>=2.17.0
